+++ R/make_group.R
@@ -4,6 +4,9 @@
 #' @param  length                A numeric.
 #' @param  n_group               A numeric.
 #' @param  tmp_group,str_length  A string to use temporary.
+#' @example
+#' unescape_utf(review) |>
+#'   make_groups(length = 1000)
 make_groups <- function(tbl, text_col = "text", length = 8000, 
                         tmp_group = "tmp_group", str_length = "str_length") {
   len_total <- sum(stringr::str_length(tbl[[text_col]]))
@@ -16,11 +19,13 @@ make_groups <- function(tbl, text_col = "text", length = 8000,
     res <- make_groups_sub(tbl, text_col, n_group, tmp_group, str_length)
     len_max <- max(res$str_length)
     if(len_max > length){
-      stop(paste0("Max length of text is over the limit\n", "lim: ", length, "\n", "max: ", len_max))
+      stop(paste0("Max length of text is over the limit\n", 
+                  "lim: ", length, "\n", "max: ", len_max))
     }
-
     len_sum_max <- max_sum_str_length(res, tmp_group, str_length)
-    if(len_sum_max < length) return(dplyr::select(res, ! dplyr::all_of("str_length")))
+    if(len_sum_max < length){
+      return(dplyr::select(res, !dplyr::all_of("str_length")))
+    }
   }
   dplyr::select(res, ! dplyr::all_of("str_length"))
 }